ai-model/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy>=2.1.0
pandas==2.2.3
scikit-learn==1.6.1
tensorflow==2.20.0
joblib==1.4.2
python-dotenv==1.0.1
gunicorn==23.0.0
